tests/requirements.txt
numpy
protobuf==3.20.3
ffmpeg-python
git+https://github.com/diggerdu/visqol-py.git
